Make diff symmetric around an item
So this is a pain in the ass, so here's how it goes:
Suppose you want to divide a stream by its differential.
Ideally, you want each item to be divided by the instantaneous
differential around itself, but doing a straight item-by-item
subtraction offsets this by half a sample.  That'll add noise
to the output stream.  So, we diff symetrically around the signal.